fix issues with mypy

+++ tests/test_timer.py
@@ -6,19 +6,19 @@ Tests for countdown timer logic in timer.py
 from focus.timer import run_countdown
 
 def test_countdown_completes() -> None:
-    ticks: list[int] = []
+    ticks: list[float] = []
     elapsed, status = run_countdown(3, lambda e, t: ticks.append(e), tick_interval=0)
     assert status == "completed"
     assert elapsed == 3
 
 def test_countdown_calls_on_tick() -> None:
-    ticks: list[int] = []
+    ticks: list[float] = []
     run_countdown(5, lambda e, t: ticks.append(e), tick_interval=0)
     assert len(ticks) == 6  # 0..5 inclusive
     assert ticks == [0, 1, 2, 3, 4, 5]
 
 def test_countdown_interrupts() -> None:
-    ticks: list[int] = []
+    ticks: list[float] = []
     # Simulate KeyboardInterrupt by raising it in the on_tick callback
     def on_tick(e: float, t: float) -> None:
         ticks.append(e)
